requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docker==6.1.3
Flask==2.3.2
Flask-RESTful==0.3.10
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psutil==5.9.5
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.3
websocket-client==1.6.1
Werkzeug==2.3.6
